+++ src/control_plane/tls/mod.rs
@@ -1,10 +1,14 @@
 #[cfg(test)]
 mod tests;
 
+mod filesystem_certificate_store;
 mod generate_local_certificates;
 mod local_certificate_bundle;
 mod local_certificate_error;
+mod stored_certificate_paths;
 
+pub(crate) use filesystem_certificate_store::FilesystemCertificateStore;
 pub(crate) use generate_local_certificates::generate_local_certificates;
 pub(crate) use local_certificate_bundle::LocalCertificateBundle;
 pub(crate) use local_certificate_error::LocalCertificateError;
+pub(crate) use stored_certificate_paths::StoredCertificatePaths;